sceance 7

+++ src/Controller/BookController.php
@@ -15,14 +15,12 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class BookController extends AbstractController
 {
-    // Page d'accueil du book controller
     #[Route('/book', name: 'app_book')]
     public function index(): Response
     {
         return $this->redirectToRoute('book_list');
     }
 
-    // Ajouter un livre
     #[Route('/book/add', name: 'addBook')]
     public function addBook(Request $request, EntityManagerInterface $em): Response
     {
@@ -51,7 +49,6 @@ class BookController extends AbstractController
         ]);
     }
 
-    // Liste des livres
     #[Route('/books', name: 'book_list')]
     public function list(BookRepository $bookRepo): Response
     {
@@ -66,10 +63,13 @@ class BookController extends AbstractController
         ]);
     }
 
-    // Modifier un livre
     #[Route('/book/edit/{id}', name: 'book_edit')]
-    public function editBook(Book $book, Request $request, EntityManagerInterface $em): Response
+    public function editBook(?Book $book, Request $request, EntityManagerInterface $em): Response
     {
+        if (!$book) {
+            throw $this->createNotFoundException('Livre non trouvé');
+        }
+
         $form = $this->createForm(BookType::class, $book);
         $form->handleRequest($request);
 
@@ -85,13 +85,17 @@ class BookController extends AbstractController
         ]);
     }
 
-    // Supprimer un livre
     #[Route('/book/delete/{id}', name: 'book_delete')]
-    public function deleteBook(Book $book, EntityManagerInterface $em): Response
+    public function deleteBook(?Book $book, EntityManagerInterface $em): Response
     {
+        if (!$book) {
+            throw $this->createNotFoundException('Livre non trouvé');
+        }
+
         $author = $book->getAuthor();
         if ($author) {
             $author->setNbBooks($author->getNbBooks() - 1);
+            $em->persist($author);
         }
 
         $em->remove($book);
@@ -101,7 +105,6 @@ class BookController extends AbstractController
         return $this->redirectToRoute('book_list');
     }
 
-    // Supprimer les auteurs sans livre
     #[Route('/authors/delete-empty', name: 'author_delete_empty')]
     public function deleteEmptyAuthors(AuthorRepository $authorRepo, EntityManagerInterface $em): Response
     {
@@ -113,5 +116,85 @@ class BookController extends AbstractController
 
         $this->addFlash('success', 'Auteurs sans livre supprimés avec succès !');
         return $this->redirectToRoute('book_list');
+    }
+
+    #[Route('/book/{id}', name: 'book_show', requirements: ['id' => '\d+'])]
+    public function show(int $id, BookRepository $bookRepo): Response
+    {
+        $book = $bookRepo->find($id);
+
+        if (!$book) {
+            throw $this->createNotFoundException('Livre non trouvé');
+        }
+
+        return $this->render('book/show.html.twig', [
+            'book' => $book,
+        ]);
+    }
+
+    #[Route('/book/search', name: 'book_search')]
+    public function search(Request $request, BookRepository $bookRepo): Response
+    {
+        $ref = $request->query->get('ref');
+        $books = [];
+
+        if ($ref) {
+            $books = $bookRepo->searchBookByRef($ref);
+        }
+
+        return $this->render('book/search_results.html.twig', [
+            'books' => $books,
+            'searchRef' => $ref,
+        ]);
+    }
+
+    #[Route('/book/update-category', name: 'book_update_category')]
+    public function updateCategory(BookRepository $bookRepo, EntityManagerInterface $em): Response
+    {
+        $affectedRows = $bookRepo->updateCategoryFromTo('Science-Fiction', 'Romance');
+
+        $this->addFlash('success', sprintf(
+            'Catégorie des livres Science-Fiction modifiée en Romance avec succès ! (%d livres modifiés)',
+            $affectedRows
+        ));
+
+        return $this->redirectToRoute('book_list');
+    }
+
+    #[Route('/books/romance/count', name: 'count_romance_books')]
+    public function countRomanceBooks(BookRepository $bookRepository): Response
+    {
+        $count = $bookRepository->countBooksByCategory('Romance');
+
+        return new Response("Nombre de livres de catégorie 'Romance' : " . $count);
+    }
+
+    #[Route('/books/published/between', name: 'books_published_between')]
+    public function booksPublishedBetween(BookRepository $bookRepository): Response
+    {
+        $start = new \DateTime('2014-01-01');
+        $end   = new \DateTime('2018-12-31');
+
+        $books = $bookRepository->findBooksPublishedBetween($start, $end);
+
+        $booksPublished = [];
+        $booksUnpublished = [];
+
+        foreach ($books as $book) {
+            if ($book->isPublished()) {
+                $booksPublished[] = $book;
+            } else {
+                $booksUnpublished[] = $book;
+            }
+        }
+
+        return $this->render('book/list.html.twig', [
+            'booksPublished' => $booksPublished,
+            'booksUnpublished' => $booksUnpublished,
+            'publishedCount' => count($booksPublished),
+            'unpublishedCount' => count($booksUnpublished),
+            'start' => $start,
+            'end' => $end,
+        ]);
     }
 }
